requirements.txt
django>=5.1.4
djangorestframework>=3.14.0,
django-cors-headers>=4.4.0
stripe>=3.0.0,<4.0
asgiref>=3.8.1 
backports.zoneinfo>=0.2.1 
sqlparse>=0.5.3 
typing-extensions>=4.12.2
